+++ l10n/de_DE.js
@@ -6,6 +6,14 @@
         'View'                                                                             : 'Anzeigen',
         'Do not warn again'                                                                : 'Nicht erneut warnen',
         'Opening HTML files in the browser can be a security risk. Do you want to proceed?': 'HTML-Dateien im Browser zu öffnen kann ein Sicherheitsrisiko sein. Möchten Sie fortfahren?',
-        'The file "{file}" ({size}) exceeds the maximum allowed size of {maxSize}.'        : 'Die Datei "{file}" ({size}) überschreitet die erlaubte Dateigröße von {maxSize}.'
+        'The file "{file}" ({size}) exceeds the maximum allowed size of {maxSize}.'        : 'Die Datei "{file}" ({size}) überschreitet die erlaubte Dateigröße von {maxSize}.',
+        '"{file}" is saved'                                                                : '"{file}" ist gespeichert',
+        'Error saving "{file}"'                                                            : 'Fehler beim Speichern von "{file}"',
+        'Saving "{file}"'                                                                  : '"{file}" wird gespeichert',
+        'Save "{file}"'                                                                    : '"{file}" speichern',
+        'Edit "{file}"'                                                                    : '"{file}" bearbeiten',
+        'Close editor'                                                                     : 'Editor schließen',
+        'Save "{file}" and close editor'                                                   : '"{file}" speichern und Editor schließen',
+        'Undo changes'                                                                     : 'Änderungen rückgängig machen'
     }, "nplurals=2; plural=(n != 1);");
 }());