+++ l10n/es.js
@@ -6,6 +6,14 @@
         'View'                                                                             : 'Mostrar',
         'Do not warn again'                                                                : 'No volver a avisar',
         'Opening HTML files in the browser can be a security risk. Do you want to proceed?': 'Abrir archivos HTML en el navegador puede ser un riesgo para la seguridad. ¿Desea continuar?',
-        'The file "{file}" ({size}) exceeds the maximum allowed size of {maxSize}.'        : 'El archivo "{file}" ({size}) supera el tamaño de archivo permitido de {maxSize}.'
+        'The file "{file}" ({size}) exceeds the maximum allowed size of {maxSize}.'        : 'El archivo "{file}" ({size}) supera el tamaño de archivo permitido de {maxSize}.',
+        '"{file}" is saved'                                                                : '"{file}" se guarda',
+        'Error saving "{file}"'                                                            : 'Error al guardar "{file}"',
+        'Saving "{file}"'                                                                  : 'Guardando "{file}"',
+        'Save "{file}"'                                                                    : 'Guardar "{file}"',
+        'Edit "{file}"'                                                                    : 'Editar "{file}"',
+        'Close editor'                                                                     : 'Cerrar editor',
+        'Save "{file}" and close editor'                                                   : 'Guardar "{file}" y cerrar el editor',
+        'Undo changes'                                                                     : 'Deshacer cambios'
     }, "nplurals=2; plural=(n != 1);");
 }());