+++ l10n/fr.js
@@ -6,6 +6,14 @@
         'View'                                                                             : 'Voir',
         'Do not warn again'                                                                : 'Ne pas réitérer l\'alerte',
         'Opening HTML files in the browser can be a security risk. Do you want to proceed?': 'Ouvrir des fichiers HTML dans le navigateur peut être un risque pour la sécurité. Vous souhaitez continuer ?',
-        'The file "{file}" ({size}) exceeds the maximum allowed size of {maxSize}.'        : 'Le fichier « {file} » ({size}) dépasse la taille de fichier autorisée de {maxSize}.'
+        'The file "{file}" ({size}) exceeds the maximum allowed size of {maxSize}.'        : 'Le fichier « {file} » ({size}) dépasse la taille de fichier autorisée de {maxSize}.',
+        '"{file}" is saved'                                                                : '"{file}" est sauvegardé',
+        'Error saving "{file}"'                                                            : 'Erreur lors de la sauvegarde de "{file}"',
+        'Saving "{file}"'                                                                  : 'Sauvegarde de "{file}"',
+        'Save "{file}"'                                                                    : 'Sauvegarder "{file}"',
+        'Edit "{file}"'                                                                    : 'Modifier "{file}"',
+        'Close editor'                                                                     : 'Fermer l\'éditeur',
+        'Save "{file}" and close editor'                                                   : 'Sauvegarder "{file}" et fermer l\'éditeur',
+        'Undo changes'                                                                     : 'Annuler les modifications'
     }, "nplurals=2; plural=(n != 1);");
 }());