+++ l10n/it.js
@@ -6,6 +6,14 @@
         'View'                                                                             : 'Visualizza',
         'Do not warn again'                                                                : 'Non avvertire di nuovo',
         'Opening HTML files in the browser can be a security risk. Do you want to proceed?': 'L\'apertura di file HTML nel browser può costituire un rischio per la sicurezza. Volete procedere?',
-        'The file "{file}" ({size}) exceeds the maximum allowed size of {maxSize}.'        : 'Il file "{file}" ({size}) supera la dimensione massima consentita di {maxSize}.'
+        'The file "{file}" ({size}) exceeds the maximum allowed size of {maxSize}.'        : 'Il file "{file}" ({size}) supera la dimensione massima consentita di {maxSize}.',
+        '"{file}" is saved'                                                                : '"{file}" viene salvato',
+        'Error saving "{file}"'                                                            : 'Errore nel salvataggio di "{file}"',
+        'Saving "{file}"'                                                                  : 'Salvataggio di "{file}"',
+        'Save "{file}"'                                                                    : 'Salvare "{file}"',
+        'Edit "{file}"'                                                                    : 'Modificare "{file}"',
+        'Close editor'                                                                     : 'Chiudere l\'editor',
+        'Save "{file}" and close editor'                                                   : 'Salvare "{file}" e chiudere l\'editor',
+        'Undo changes'                                                                     : 'Annullare le modifiche'
     }, "nplurals=2; plural=(n != 1);");
 }());