+++ src/HTMLViewer.vue
@@ -9,8 +9,9 @@
   -->
 
 <template>
-    <div class="htmlviewer-container">
-        <ContentContainer :file="file" :show-name="isComparison" v-if="file.loaded"/>
+    <div class="htmlviewer-container" :class="{'htmlviewer-has-editor': isEdit}">
+        <ContentEditor :file="file" v-if="file.loaded && this.isEdit"/>
+        <ContentViewer :file="file" :show-name="isComparison" :show-edit="showEditButton" v-on:edit="isEdit=true;" v-if="file.loaded"/>
         <ErrorMessage :file="file" v-if="file.isTooBig"/>
         <WarningDialog v-on:abort="abort" v-on:confirm="ok" v-if="hasDialog"/>
         <NcLoadingIcon class="htmlviewer-loader" :size="64" v-if="showLoading && !file.loaded"/>
@@ -24,13 +25,20 @@
     import scriptWarning from "./utils/scriptWarning";
     import ErrorMessage from "./components/ErrorMessage.vue";
     import WarningDialog from "./components/WarningDialog.vue";
-    import ContentContainer from "./components/ContentContainer.vue";
+    import ContentViewer from "./components/ContentViewer.vue";
 
     export default {
-        components: {ContentContainer, WarningDialog, ErrorMessage, NcLoadingIcon},
+        components: {
+            ContentViewer,
+            ContentEditor: () => import/* webpackChunkName: "ContentEditor" */("./components/ContentEditor.vue"),
+            WarningDialog,
+            ErrorMessage,
+            NcLoadingIcon
+        },
         props     : {
             size       : Number,
             source     : String,
+            permissions: String,
             showLoading: {
                 type   : Boolean,
                 default: false
@@ -45,6 +53,7 @@
 
             return {
                 file,
+                isEdit   : false,
                 hasDialog: false
             };
         },
@@ -59,6 +68,9 @@
         computed: {
             isComparison() {
                 return this.$parent.$el.querySelector('.viewer__content.viewer--split') !== null;
+            },
+            showEditButton() {
+                return !this.isComparison && !this.isEdit && this.file.loaded && this.permissions.indexOf('W') !== -1;
             }
         },
         methods : {
@@ -89,7 +101,8 @@
 
         watch: {
             source(value) {
-                this.file = new File(this.basename, value, this.size);
+                this.file = new File(this.basename, value, this.size, this.fileVersion);
+                this.isEdit = false;
                 this.loadFile().catch(console.error);
             }
         }
@@ -100,6 +113,16 @@
 .htmlviewer-container {
     width  : 100% !important;
     height : 100%;
+
+    &.htmlviewer-has-editor {
+        display : flex;
+        gap     : .25rem;
+
+        .htmlviewer-content {
+            border-radius : var(--border-radius-large);
+            overflow      : hidden;
+        }
+    }
 
     .htmlviewer-loader {
         margin-top : 5rem;